// src/interfaces/global.ts
export interface IModalTechData {
  title: string;
  desc: string;
}

export interface IServiceItems {
  title: string;
  desc: string;
}
export interface IFormData {
  name: string;
  email: string;
  msg: string;
}

export interface IFormDataResponse {
  nameValue: string;
  emailValue: string;
  msgValue: string;
}